# docs/openapi.yaml
openapi: 3.0.3
info:
  title: IndianTradeMart API
  version: 1.0.0
  description: Express API surface for the IndianTradeMart platform.
servers:
  - url: http://localhost:3001
    description: Local dev server
  - url: / 
    description: Relative (proxy)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PasswordResetRequest:
      type: object
      required: [email, new_password]
      properties:
        email:
          type: string
          format: email
        new_password:
          type: string
          minLength: 6
...
    ChatRequest:
      type: object
      required: [messages]
      properties:
        language:
          type: string
        model:
          type: string
        messages:
          type: array
          items:
            type: object
            required: [role, text]
            properties:
              role:
                type: string
              text:
                type: string
...
paths:
...
  /api/password-reset:
    post:
      summary: Reset password after OTP
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
      responses:
        '200':
          description: Password updated
...
  /api/admin/vendors:
    get:
      summary: Admin list vendors
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Vendors
...
  /api/chat:
    post:
      summary: Chatbot
      tags: [Chatbot]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: Reply
...
  /api/superadmin/vendors:
    get:
      summary: Superadmin list vendors
      tags: [Superadmin]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Vendors
...